Fix login and fix warnings
